Move delete confirmation to separate file and add permissions

+++ app/views/newsletter.py
@@ -1,7 +1,9 @@
 from flask import Blueprint, request, render_template, redirect, url_for,\
-    flash, Response, render_template_string
+    flash, Response, abort
 from flask_babel import _
+
 from app import db
+from app.utils.module import ModuleAPI
 from app.forms.newsletter import NewsletterForm
 from app.models.newsletter import Newsletter
 
@@ -10,6 +12,9 @@ blueprint = Blueprint('newsletter', __name__, url_prefix='/newsletter')
 
 @blueprint.route('/', methods=['GET'])
 def all():
+    if not ModuleAPI.can_read('newsletter'):
+        return abort(403)
+
     newsletters = Newsletter.query.all()
     return render_template('newsletter/view.htm', newsletters=newsletters)
 
@@ -17,7 +22,9 @@ def all():
 @blueprint.route('/create/', methods=['GET', 'POST'])
 @blueprint.route('/edit/<int:newsletter_id>/', methods=['GET', 'POST'])
 def edit(newsletter_id=None):
-    # TODO check for admin
+    if not ModuleAPI.can_write('newsletter'):
+        return abort(403)
+
     if newsletter_id:
         newsletter = Newsletter.query.get_or_404(newsletter_id)
     else:
@@ -38,26 +45,11 @@ def edit(newsletter_id=None):
 
 @blueprint.route('/delete/<int:newsletter_id>/', methods=['GET', 'POST'])
 def delete(newsletter_id):
-    # TODO check for admin
+    if not ModuleAPI.can_write('newsletter'):
+        return abort(403)
 
-    # TODO make this its own template
     if request.method == 'GET':
-        return render_template_string('''
-            {% extends "content.htm" %}
-
-            {% block content %}
-            <h1>{{ _('Delete') }}</h1>
-            <h2>{{ _('Are you sure you want to delete this entry?') }}</h2>
-            <form method="POST">
-                <input type="submit" class="btn btn-danger"
-                       value="{{ _('Yes') }}">
-
-                <a href="{{ url_for('.all') }}" class="btn btn-primary">
-                    {{ _('No') }}
-                </a>
-            </form>
-            {% endblock %}
-        ''')
+        return render_template('newsletter/confirm.htm')
 
     newsletter = Newsletter.query.get_or_404(newsletter_id)
     db.session.delete(newsletter)
@@ -68,6 +60,9 @@ def delete(newsletter_id):
 
 @blueprint.route('/<int:newsletter_id>/activities/', methods=['GET'])
 def activities_xml(newsletter_id):
+    if not ModuleAPI.can_read('newsletter'):
+        return abort(403)
+
     newsletter = Newsletter.query.get_or_404(newsletter_id)
     headers = {'Content-disposition': 'attachment; filename=activities.xml'}
     return Response(render_template('newsletter/activities.xml',
@@ -77,6 +72,9 @@ def activities_xml(newsletter_id):
 
 @blueprint.route('/<int:newsletter_id>/news/', methods=['GET'])
 def news_xml(newsletter_id):
+    if not ModuleAPI.can_read('newsletter'):
+        return abort(403)
+
     newsletter = Newsletter.query.get_or_404(newsletter_id)
     headers = {'Content-disposition': 'attachment; filename=news.xml'}
     return Response(render_template('newsletter/news.xml',
